Move MembersSelector
