+++ app/src/components/MembersSelector/MembersSelector.scss
@@ -1,5 +1,5 @@
-@import '../../../../variables.scss';
-@import '../../../../styles.scss';
+@import '../../variables.scss';
+@import '../../styles.scss';
 
 .members-selector {
   .table {
