+++ app/src/pages/CreateGroup/CreateGroup.jsx
@@ -6,8 +6,8 @@ import { Helmet } from 'react-helmet';
 import PageTitle from '../../components/PageTitle';
 import TextInput from '../../components/TextInput';
 import TextButton from '../../components/TextButton';
+import MembersSelector from '../../components/MembersSelector';
 import Button from '../../components/Button';
-import MembersSelector from './components/MembersSelector';
 import ImportPlayersModal from '../../modals/ImportPlayersModal';
 import VerificationModal from '../../modals/VerificationModal';
 import createGroupAction from '../../redux/modules/groups/actions/create';
