--- app/src/pages/EditGroup/components/MembersSelector/MembersSelector.scss
@@ -1,36 +1,0 @@
-@import '../../../../variables.scss';
-@import '../../../../styles.scss';
-
-.members-selector {
-  .table {
-    margin-top: 30px;
-
-    .table-btn {
-      background: none;
-      border: none;
-      padding: 0;
-      cursor: pointer;
-
-      &:active {
-        opacity: 0.5;
-      }
-
-      &.-switch-role {
-        color: $neutral-orange;
-      }
-
-      &.-remove {
-        color: $negative-red;
-      }
-    }
-  }
-
-  .empty-selected {
-    @extend %panel;
-    width: 100%;
-    display: block;
-    padding: 15px;
-    font-size: 0.9em;
-    margin: 20px 0;
-  }
-}
