--- app/src/pages/EditGroup/components/MembersSelector/index.js
@@ -1,3 +1,0 @@
-import MembersSelector from './MembersSelector';
-
-export default MembersSelector;
